+++ Runtime/Component/Sub/TextStyle.cs
@@ -1,4 +1,3 @@
-using System.Threading;
 /******************************************/
 /*                                        */
 /*     Copyright (c) 2018 monitor1394     */
@@ -18,11 +17,18 @@ namespace XCharts
     [Serializable]
     public class TextStyle : SubComponent, IEquatable<TextStyle>
     {
+        [SerializeField] private Font m_Font;
         [SerializeField] private float m_Rotate = 0;
         [SerializeField] private Vector2 m_Offset = Vector2.zero;
         [SerializeField] private Color m_Color = Color.clear;
+        [SerializeField] private Color m_BackgroundColor = Color.clear;
         [SerializeField] private int m_FontSize = 18;
         [SerializeField] private FontStyle m_FontStyle = FontStyle.Normal;
+        [SerializeField] private float m_LineSpacing = 1f;
+        [SerializeField] private float m_PaddingLeft = 0f;
+        [SerializeField] private float m_PaddingRight = 0f;
+        [SerializeField] private float m_PaddingTop = 0f;
+        [SerializeField] private float m_PaddingBottom = 0f;
 
         /// <summary>
         /// Rotation of text.
@@ -43,6 +49,16 @@ namespace XCharts
         /// </summary>
         public Color color { get { return m_Color; } set { m_Color = value; } }
         /// <summary>
+        /// the color of text. 
+        /// 文本的背景颜色。
+        /// </summary>
+        public Color backgroundColor { get { return m_BackgroundColor; } set { m_BackgroundColor = value; } }
+        /// <summary>
+        /// the font of text.
+        /// 文本字体
+        /// </summary>
+        public Font font { get { return m_Font; } set { m_Font = value; } }
+        /// <summary>
         /// font size.
         /// 文本字体大小。
         /// </summary>
@@ -52,6 +68,11 @@ namespace XCharts
         /// 文本字体的风格。
         /// </summary>
         public FontStyle fontStyle { get { return m_FontStyle; } set { m_FontStyle = value; } }
+        /// <summary>
+        /// text line spacing.
+        /// 行间距。
+        /// </summary>
+        public float lineSpacing { get { return m_LineSpacing; } set { m_LineSpacing = value; } }
 
         public TextStyle()
         {
@@ -88,8 +109,10 @@ namespace XCharts
             this.fontSize = style.fontSize;
             this.fontStyle = style.fontStyle;
             this.color = style.color;
+            this.backgroundColor = style.backgroundColor;
             this.rotate = style.rotate;
             this.offset = style.offset;
+            this.lineSpacing = style.lineSpacing;
         }
 
         public TextStyle Clone()
@@ -97,9 +120,11 @@ namespace XCharts
             var textStyle = new TextStyle();
             textStyle.rotate = rotate;
             textStyle.color = color;
+            textStyle.backgroundColor = backgroundColor;
             textStyle.fontSize = fontSize;
             textStyle.fontStyle = fontStyle;
             textStyle.offset = offset;
+            textStyle.lineSpacing = lineSpacing;
             return textStyle;
         }
 
@@ -129,6 +154,8 @@ namespace XCharts
                 fontSize == other.fontSize &&
                 fontStyle == other.fontStyle &&
                 offset == other.offset &&
+                lineSpacing == other.lineSpacing &&
+                ChartHelper.IsValueEqualsColor(m_BackgroundColor, other.backgroundColor) &&
                 ChartHelper.IsValueEqualsColor(m_Color, other.color);
         }